+++ server/routes/instance/structure.ts
@@ -26,6 +26,7 @@ import { getGlobalAdmin, getGlobalNonAdmin } from "../../project_auth.ts";
 import { defineRoute } from "../route-helpers.ts";
 import { streamResponse } from "../streaming.ts";
 import { log } from "../../middleware/logging.ts";
+import { requireGlobalPermission } from "../../middleware/userPermission.ts";
 
 export const routesStructure = new Hono();
 
@@ -38,7 +39,7 @@ export const routesStructure = new Hono();
 defineRoute(
   routesStructure,
   "getStructureItems",
-  getGlobalNonAdmin,
+  requireGlobalPermission("can_view_data"),
   log("getStructureItems"),
   async (c) => {
     const res = await getStructureItems(c.var.mainDb, _DATASET_LIMIT);
@@ -49,7 +50,7 @@ defineRoute(
 defineRoute(
   routesStructure,
   "deleteAllStructureData",
-  getGlobalAdmin,
+  requireGlobalPermission("can_configure_data"),
   log("deleteAllStructureData"),
   async (c) => {
     const res = await deleteAllStructureData(c.var.mainDb);
@@ -66,7 +67,7 @@ defineRoute(
 defineRoute(
   routesStructure,
   "addStructureUploadAttempt",
-  getGlobalAdmin,
+  requireGlobalPermission("can_configure_data"),
   log("addStructureUploadAttempt"),
   async (c) => {
     const res = await addStructureUploadAttempt(c.var.mainDb);
@@ -77,7 +78,7 @@ defineRoute(
 defineRoute(
   routesStructure,
   "getStructureUploadAttempt",
-  getGlobalAdmin,
+  requireGlobalPermission("can_configure_data"),
   log("getStructureUploadAttempt"),
   async (c) => {
     const res = await getStructureUploadAttempt(c.var.mainDb);
@@ -88,7 +89,7 @@ defineRoute(
 defineRoute(
   routesStructure,
   "deleteStructureUploadAttempt",
-  getGlobalAdmin,
+  requireGlobalPermission("can_configure_data"),
   log("deleteStructureUploadAttempt"),
   async (c) => {
     const res = await deleteStructureUploadAttempt(c.var.mainDb);
@@ -105,7 +106,7 @@ defineRoute(
 defineRoute(
   routesStructure,
   "structureStep1Csv_UploadFile",
-  getGlobalAdmin,
+  requireGlobalPermission("can_configure_data"),
   log("structureStep1Csv_UploadFile"),
   async (c, { body }) => {
     const res = await structureStep1Csv_UploadFile(
@@ -119,7 +120,7 @@ defineRoute(
 defineRoute(
   routesStructure,
   "structureStep2Csv_SetColumnMappings",
-  getGlobalAdmin,
+  requireGlobalPermission("can_configure_data"),
   log("structureStep2Csv_SetColumnMappings"),
   async (c, { body }) => {
     const res = await structureStep2Csv_SetColumnMappings(
@@ -133,7 +134,7 @@ defineRoute(
 defineRoute(
   routesStructure,
   "structureStep3Csv_StageData",
-  getGlobalAdmin,
+  requireGlobalPermission("can_configure_data"),
   log("structureStep3Csv_StageData"),
   async (c) => {
     const res = await structureStep3Csv_StageData(c.var.mainDb);
@@ -144,7 +145,7 @@ defineRoute(
 defineRoute(
   routesStructure,
   "structureStep3Csv_StageDataStreaming",
-  getGlobalAdmin,
+  requireGlobalPermission("can_configure_data"),
   log("structureStep3Csv_StageDataStreaming"),
   (c) => {
     return streamResponse(c, async (writer) => {
@@ -173,7 +174,7 @@ defineRoute(
 defineRoute(
   routesStructure,
   "structureStep3Dhis2_StageData",
-  getGlobalAdmin,
+  requireGlobalPermission("can_configure_data"),
   log("structureStep3Dhis2_StageData"),
   async (c) => {
     const res = await structureStep3Dhis2_StageData(c.var.mainDb);
@@ -184,7 +185,7 @@ defineRoute(
 defineRoute(
   routesStructure,
   "structureStep3Dhis2_StageDataStreaming",
-  getGlobalAdmin,
+  requireGlobalPermission("can_configure_data"),
   log("structureStep3Dhis2_StageDataStreaming"),
   (c) => {
     return streamResponse(c, async (writer) => {
@@ -213,7 +214,7 @@ defineRoute(
 defineRoute(
   routesStructure,
   "structureStep4_ImportData",
-  getGlobalAdmin,
+  requireGlobalPermission("can_configure_data"),
   log("structureStep4_ImportData"),
   async (c, { body }) => {
     const res = await structureStep4_ImportData(c.var.mainDb, body.strategy);
@@ -224,7 +225,7 @@ defineRoute(
 defineRoute(
   routesStructure,
   "structureStep0_SetSourceType",
-  getGlobalAdmin,
+  requireGlobalPermission("can_configure_data"),
   log("structureStep0_SetSourceType"),
   async (c, { body }) => {
     const res = await structureStep0_SetSourceType(
@@ -238,7 +239,7 @@ defineRoute(
 defineRoute(
   routesStructure,
   "structureStep1Dhis2_SetCredentials",
-  getGlobalAdmin,
+  requireGlobalPermission("can_configure_data"),
   log("structureStep1Dhis2_SetCredentials"),
   async (c, { body }) => {
     // Validate credentials with DHIS2 first
@@ -261,7 +262,7 @@ defineRoute(
 defineRoute(
   routesStructure,
   "structureStep2Dhis2_SetOrgUnitSelection",
-  getGlobalAdmin,
+  requireGlobalPermission("can_configure_data"),
   log("structureStep2Dhis2_SetOrgUnitSelection"),
   async (c, { body }) => {
     const res = await structureStep2Dhis2_SetOrgUnitSelection(
@@ -275,7 +276,7 @@ defineRoute(
 defineRoute(
   routesStructure,
   "getStructureUploadStatus",
-  getGlobalAdmin,
+  requireGlobalPermission("can_configure_data"),
   log("getStructureUploadStatus"),
   async (c) => {
     const res = await getStructureUploadStatus(c.var.mainDb);
@@ -286,7 +287,7 @@ defineRoute(
 defineRoute(
   routesStructure,
   "structureStep2Dhis2_GetOrgUnitsMetadata",
-  getGlobalAdmin,
+  requireGlobalPermission("can_configure_data"),
   log("structureStep2Dhis2_GetOrgUnitsMetadata"),
   async (c) => {
     // Get the current upload attempt
@@ -331,7 +332,7 @@ defineRoute(
 // CSV export endpoint - uses getStructureItems without limit for all rows
 routesStructure.get(
   "/structure/facilities/export/csv",
-  getGlobalNonAdmin,
+  requireGlobalPermission("can_configure_data"),
   log("exportStructureItemsCsv"),
   async (c) => {
     // Get all facilities with proper column selection based on maxAdminArea
